propagate cargo features to ir expansion
